requirements.txt
backoff
certifi
charset-normalizer
colorama
idna
opencage
phonenumbers
python-dotenv
requests
urllib3
